+++ src/goto-programs/adjust_float_expressions.cpp
@@ -11,12 +11,13 @@ Author: Daniel Kroening, kroening@kroening.com
 
 #include "adjust_float_expressions.h"
 
+#include <util/arith_tools.h>
 #include <util/cprover_prefix.h>
 #include <util/expr_util.h>
+#include <util/floatbv_expr.h>
+#include <util/ieee_float.h>
 #include <util/std_expr.h>
 #include <util/symbol.h>
-#include <util/ieee_float.h>
-#include <util/arith_tools.h>
 
 #include "goto_model.h"
