+++ src/solvers/flattening/boolbv_floatbv_op.cpp
@@ -11,6 +11,7 @@ Author: Daniel Kroening, kroening@kroening.com
 #include <algorithm>
 #include <iostream>
 
+#include <util/floatbv_expr.h>
 #include <util/std_types.h>
 
 #include <solvers/floatbv/float_utils.h>
