+++ src/solvers/floatbv/float_bv.cpp
@@ -12,6 +12,7 @@ Author: Daniel Kroening, kroening@kroening.com
 
 #include <util/arith_tools.h>
 #include <util/bitvector_expr.h>
+#include <util/floatbv_expr.h>
 #include <util/std_expr.h>
 
 exprt float_bvt::convert(const exprt &expr) const
