+++ src/solvers/refinement/refine_arithmetic.cpp
@@ -8,10 +8,11 @@ Author: Daniel Kroening, kroening@kroening.com
 
 #include "bv_refinement.h"
 
-#include <util/bv_arithmetic.h>
-#include <util/ieee_float.h>
-#include <util/expr_util.h>
 #include <util/arith_tools.h>
+#include <util/bv_arithmetic.h>
+#include <util/expr_util.h>
+#include <util/floatbv_expr.h>
+#include <util/ieee_float.h>
 
 #include <solvers/floatbv/float_utils.h>
