+++ src/util/simplify_expr_floatbv.cpp
@@ -11,6 +11,7 @@ Author: Daniel Kroening, kroening@kroening.com
 #include "arith_tools.h"
 #include "expr.h"
 #include "expr_util.h"
+#include "floatbv_expr.h"
 #include "ieee_float.h"
 #include "invariant.h"
 #include "namespace.h"
